requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.14.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-translate==3.13.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.48.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
mailjet-rest==1.3.4
MarkupSafe==2.1.3
packaging==23.2
protobuf==3.20.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.2
xpinyin==0.7.6
